agent-license-once-compare: reanalyze no longer complains about an unlink problem.
git-svn-id: http://svn.code.sf.net/p/fossology/code/trunk@1449 a5643f74-2041-0410-a2b9-94ffef304a07

+++ fossology/ui/plugins/agent-license-once-compare.php
@@ -91,7 +91,7 @@ class agent_license_once_compare extends FO_Plugin
     $V .= $OneShot->AnalyzeOne($Highlight,$TempLics);
 
     /* Clean up */
-    unlink($TempCache);
+    if (file_exists($TempCache)) { unlink($TempCache); }
     return($V);
   } // AnalyzeOne()
